Add controlled form component
